# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/activities": {
      "get": {
        "operationId": "ActivitiesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Activities"
        ]
      },
      "post": {
        "operationId": "ActivitiesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/v1/activities/{id}": {
      "get": {
        "operationId": "ActivitiesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Activities"
        ]
      },
      "patch": {
        "operationId": "ActivitiesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/v1/activities/{id}/status": {
      "patch": {
        "operationId": "ActivitiesController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/v1/activities/{id}/roles": {
      "get": {
        "operationId": "ActivitiesController_roles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Activities"
        ]
      },
      "post": {
        "operationId": "ActivitiesController_createRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/v1/activities/{id}/roles/{roleId}/applications": {
      "post": {
        "operationId": "ActivitiesController_apply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/v1/activities/{id}/applications": {
      "get": {
        "operationId": "ActivitiesController_applications",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/v1/activities/{id}/applications/{applicationId}": {
      "patch": {
        "operationId": "ActivitiesController_respond",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/v1/activities/{id}/registrations": {
      "post": {
        "operationId": "ActivitiesController_register",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/v1/activities/{id}/registrations/me": {
      "delete": {
        "operationId": "ActivitiesController_cancelRegistration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/v1/activities/{id}/participation": {
      "get": {
        "operationId": "ActivitiesController_participation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/v1/activities/{id}/evaluations": {
      "get": {
        "operationId": "ActivitiesController_evaluations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Activities"
        ]
      },
      "post": {
        "operationId": "ActivitiesController_evaluate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvaluationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/v1/me/activities": {
      "get": {
        "operationId": "ActivitiesMeController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ActivitiesMe"
        ]
      }
    },
    "/api/v1/me/organized-activities": {
      "get": {
        "operationId": "ActivitiesMeController_organized",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ActivitiesMe"
        ]
      }
    },
    "/api/v1/me/activity-capabilities": {
      "get": {
        "operationId": "ActivitiesMeController_capabilities",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ActivitiesMe"
        ]
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "MeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ]
      }
    },
    "/auth/callback": {
      "get": {
        "operationId": "SsoCallbackController_callback",
        "parameters": [
          {
            "name": "access_token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token_type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Bearer",
                "bearer"
              ]
            }
          },
          {
            "name": "expires_in",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 32,
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "accept",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SsoCallback"
        ]
      }
    },
    "/api/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    }
  },
  "info": {
    "title": "Department Activity Planner API",
    "description": "CSMJU2030 subsystem. Every success response is wrapped as { success, data[, meta] }; errors as { success: false, error: { code, message, details } }.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateActivityDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "category": {
            "type": "string",
            "enum": [
              "workshop",
              "seminar",
              "competition",
              "volunteer",
              "recreation",
              "academic",
              "other"
            ]
          },
          "location": {
            "type": "string",
            "maxLength": 200
          },
          "maxParticipants": {
            "type": "number",
            "minimum": 1,
            "maximum": 2147483647
          },
          "startAt": {
            "type": "string"
          },
          "endAt": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "category",
          "maxParticipants",
          "startAt",
          "endAt"
        ]
      },
      "UpdateActivityDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "category": {
            "type": "string",
            "enum": [
              "workshop",
              "seminar",
              "competition",
              "volunteer",
              "recreation",
              "academic",
              "other"
            ]
          },
          "location": {
            "type": "string",
            "maxLength": 200
          },
          "maxParticipants": {
            "type": "number",
            "minimum": 1,
            "maximum": 2147483647
          },
          "startAt": {
            "type": "string"
          },
          "endAt": {
            "type": "string"
          }
        }
      },
      "UpdateActivityStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "FULL",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateActivityRoleDto": {
        "type": "object",
        "properties": {
          "roleName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "maxMembers": {
            "type": "number",
            "minimum": 1,
            "maximum": 100
          }
        },
        "required": [
          "roleName",
          "maxMembers"
        ]
      },
      "RespondApplicationDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACCEPTED",
              "REJECTED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateEvaluationDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "liked": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "improvement": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "rating",
          "liked",
          "improvement"
        ]
      }
    }
  }
}